add targeting button and refactor getUnitInfos into api layer [Re: nmkataoka/adv-life#28]
