+++ .eslintrc
@@ -33,6 +33,7 @@
     "react-hooks/exhaustive-deps": "warn",
     "semi": 2,
     "@typescript-eslint/explicit-function-return-type": 0,
+    "@typescript-eslint/no-unsafe-assignment": 0,
     "@typescript-eslint/indent": [
       "error",
       2,
